# [Merged by Bors] - logging: add array/object logging for ballot