## Motivation

zap provides interface for logging objects and arrays lazily and in general more efficiently

```
logger.go:130: 2021-12-01T07:20:54.920+0200	DEBUG	processing ballot	{"rerun": true, "last-layer-id": 15, "processing_ballot_id": "9c5c877751", "processing_ballot_layer": 23, "ballot": {"id": "9c5c877751", "layer": 23, "epoch": 5, "smesher": "ab5159b7cff1665ee0060a73886b6f4aa4c5a80cdd8b901cdad96e7361be0b6c", "base_ballot": "1f168bb954", "supports": 10, "againsts": 0, "abstains": 0, "atx": "1ac8d6170f", "eligibility_counter": 0, "ref_ballot": "0000000000", "active_set_size": 15, "beacon": "c5d23f6f3f"}, "name": ""}
    logger.go:130: 2021-12-01T07:20:54.920+0200	DEBUG	ballot adds support for	{"rerun": true, "last-layer-id": 15, "processing_ballot_id": "9c5c877751", "processing_ballot_layer": 23, "count": 10, "block_ids": ["60266206d3", "b961036c75", "9371a9aa73", "4d1b1fe580", "9731751652", "284d6180a6", "6c5e2ea7c3", "d81444af73", "1f168bb954", "7b26f873e1"], "name": ""}
```

for example reduction in allocations is significant in this case:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkTortoiseLayerHandling-16     900988496     836235873     -7.19%

benchmark                             old allocs     new allocs     delta
BenchmarkTortoiseLayerHandling-16     2612684        1656383        -36.60%

benchmark                             old bytes     new bytes     delta
BenchmarkTortoiseLayerHandling-16     507506576     420751108     -17.09%
```

related: https://github.com/spacemeshos/go-spacemesh/issues/2962

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)